Made adjustments to header, register, and a few other styling. Deleted Login page as well.
