--- client/src/components/Header/Header.css
@@ -1,6 +1,0 @@
-.das-navbar {
-    position: fixed !important;
-    top: 0;
-    left: 0;
-    width: 100%;
-}
