--- client/src/components/Header/HeaderLP.css
@@ -1,10 +1,0 @@
-/* .das-navbar {
-    display: flex;
-    position: fixed !important;
-    top: 0;
-    left: 0;
-    width: 100%;
-}
-.Get-Started_Btn{
-    color:rgb(184, 32, 32);
-  } */
