+++ client/src/index.css
@@ -179,6 +179,7 @@ body {
 
 /*register page*/
 
+
 input[type = "submit-btn"]{
   background-color: var(--primary-color);
   color: rgb(66, 66, 66);
@@ -198,12 +199,22 @@ input[type = "submit-btn"]{
   
 }
 
+/* Treat this bottom as your new background for register ONLY  */
 .register-page{
 
-  background-color: #2c2c2c;
+  /* background-color: #2c2c2c;
   align-items: center;
   padding-bottom: 0;
   text-align: center;
+  margin: 0 auto; */
+
+  position: absolute;
+  background-color: #2c2c2c;
+  align-items: center;
+  text-align: center;
+  margin: 0 auto;
+  width: 100%;
+  height: 100%;
   
 }
 .register-page h1{
@@ -229,16 +240,27 @@ input{
 
 }
 .register-page p{
-  padding-bottom:10em;
+  /* padding-bottom:10em; */
   color:white;
 }
+
+.register-page h2 {
+  color: white;
+}
+
 a{
   padding-bottom:0em;
   color:rgb(246, 33, 33);
   font-size: 20px;;
 }
+
+.register-page .logout {
+  float: none;
+}
+
 /*login page*/
-.sign-in-h{
+
+/* .sign-in-h{
   background-color: #161616;
   align-items: center;
 }
@@ -257,5 +279,6 @@ margin:3em;
 }
 .logo-signin{
   
-}
+} */
+
 /* Footer */
